techDebt(registrationRefactor) Renamed responseErrors filename to registrationErrors
